Moving settings and property into separate traits.
