+++ src/Traits/MorphModel/WithProperties.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace EricDowell\ResourceController\Traits\MorphModel;
+
+trait WithProperties
+{
+}
